refactor(library/Sequence): moves "CompatibilityError.ts" to "Compatibility/Error.ts"
