+++ src/library/Sequence/Byte/Encoded/Compatibility/Error.ts
